requirements.txt
streamlit
langchain
groq
python-dotenv
docx2txt
PyMuPDF
easyocr
Pillow

